withdraw funds email

+++ database/migrations/2018_06_23_093944_create_hp_commissions_table.php
@@ -16,7 +16,7 @@ class CreateHpCommissionsTable extends Migration
         Schema::create('hp_commissions', function (Blueprint $table) {
             $table->increments('id');
             $table->string('transaction_id');
-            $table->integer('commission_rate');
+            $table->float('commission_rate', 8,2);
             $table->date('commission_date');
             $table->float('commission_amount', 8,2);
             $table->text('description');
